fix long

+++ mingpt/model.py
@@ -216,7 +216,7 @@ class GPT(nn.Module):
         # if we are given some desired targets also calculate the loss
         loss = None
         if targets is not None:
-            loss = F.cross_entropy(logits.view(-1, logits.size(-1)), targets.view(-1))
+            loss = F.cross_entropy(logits.view(-1, logits.size(-1)), targets.long().view(-1))
 
             if self.bert:
                 IM = 1.0 - M
